+++ packages/core-bridge/src/conversions.rs
@@ -63,8 +63,8 @@ impl ObjectHandleConversionsExt for Handle<'_, JsObject> {
         let span_id = js_value_getter!(ctx, self, "spanId", JsString);
         let trace_flags = js_value_getter!(ctx, self, "traceFlags", JsNumber);
         Ok(SpanContext::new(
-            TraceId::from_hex(&trace_id),
-            SpanId::from_hex(&span_id),
+            TraceId::from_hex(&trace_id).expect("TraceId is valid"),
+            SpanId::from_hex(&span_id).expect("SpanId is valid"),
             TraceFlags::new(trace_flags as u8),
             false,
             TraceState::from_str("").expect("Trace state must be valid"),